Fix another bug in fingerprinting FP2, #224
When we reach an atom which is connected back the first atom in ring,
it was closing the ring using the bond order of the bond it used
to reach this last atom, not the bond-order of the bond connecting
the last atom to the first one.

I don't think this bug is path dependent, so any representation of the
same molecule should yield the same fingerprint, they will all be wrong
together.

+++ src/formats/gaussformat.cpp
@@ -332,7 +332,8 @@ namespace OpenBabel
     const char *pg = obPG.IdentifyPointGroup();
     
     double Rgas = 1.9872041; // cal/mol K http://en.wikipedia.org/wiki/Gas_constant
-    double Srot = -Rgas * log(RotSymNum);
+    double Srot = -Rgas * log(double(RotSymNum));
+
     
     //printf("DHf(M,0) = %g, DHf(M,T) = %g, S0(M,T) = %g\nPoint group = %s RotSymNum = %d Srot = %g\n",
     //       dhofM0, dhofMT, S0MT, pg, RotSymNum, Srot);
